Read inputed floats into list
Now the floats which are read from the input file are added to the
linked list of data.

* main.c - Input floats from the input file into the linked list.

* list.h - Changed list type from int to float.

* list.c - Changed print format string from %i to %f.

+++ list.h
@@ -1,7 +1,7 @@
 #ifndef _LIST_H_
 #define _LIST_H_
 
-typedef int listType;
+typedef float listType;
 
 struct _listNode {
 	listType data;
